+++ apps/src/idelab/IdelabEditor.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import $ from 'jquery';
 import AceEditor from 'react-ace';
 import {connect} from 'react-redux';
-import {setEditorText} from './javaIdeRedux';
+import {setEditorText} from './idelabRedux';
 import PropTypes from 'prop-types';
 import PaneHeader, {PaneSection} from '@cdo/apps/templates/PaneHeader';
 import 'ace-builds/src-noconflict/mode-java';
@@ -15,7 +15,7 @@ const style = {
   }
 };
 
-class JavaEditor extends React.Component {
+class IdelabEditor extends React.Component {
   static propTypes = {
     style: PropTypes.object,
     // populated by redux
@@ -45,7 +45,7 @@ class JavaEditor extends React.Component {
           mode="java"
           theme="monokai"
           onChange={this.onChange}
-          name="java-lab-editor"
+          name="ide-lab-editor"
           editorProps={{$blockScrolling: true}}
           style={style.editor}
         />
@@ -56,9 +56,9 @@ class JavaEditor extends React.Component {
 
 export default connect(
   state => ({
-    editorText: state.javaIde.editorText
+    editorText: state.idelab.editorText
   }),
   dispatch => ({
     setEditorText: editorText => dispatch(setEditorText(editorText))
   })
-)(JavaEditor);
+)(IdelabEditor);
