+++ apps/style/idelab/style.scss
