+++ dashboard/app/controllers/idelab_controller.rb
@@ -1,0 +1,8 @@
+class IdelabController < ApplicationController
+  before_action :authenticate_user!
+  load_and_authorize_resource class: :idelab
+
+  # GET /idelab
+  def index
+  end
+end
