--- dashboard/app/views/java_ide/index.html.haml
@@ -1,6 +1,0 @@
-- content_for(:head) do
-  %script{src: webpack_asset_path('js/java_ide/index.js'), data: @script_data}
-  = stylesheet_link_tag 'css/javaIde', media: 'all'
-
-#java-ide-container
-  -# populated by React
